backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
Django==3.2.15
django-extra-fields==3.0.2
django-templated-mail==1.1.1
django-filter==22.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
#drf-extra-fields==3.7.0
filetype==1.2.0
pillow==10.3.0
#psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
sqlparse==0.5.0
tzdata==2024.1
